+++ libs/gl/shader.cpp
@@ -52,7 +52,7 @@ void shader::compile( void )
 	{
 		GLint len = 0;
 		glGetShaderiv( _shader, GL_SHADER_SOURCE_LENGTH, &len );
-		std::string source( len, '\0' );
+		std::string source( static_cast<size_t>( len ), '\0' );
 		glGetShaderSource( _shader, len, nullptr, &source[0] );
 		std::cerr << "Compile error:\n" << source << std::endl;
 		throw std::runtime_error( log() );
@@ -64,7 +64,7 @@ void shader::compile( void )
 		{
 			GLint len = 0;
 			glGetShaderiv( _shader, GL_SHADER_SOURCE_LENGTH, &len );
-			std::string source( len+1, '\0' );
+			std::string source( static_cast<size_t>( len + 1 ), '\0' );
 			glGetShaderSource( _shader, len+1, nullptr, &source[0] );
 			std::cerr << "========================================" << std::endl;
 			std::cerr << "Compile warning:\n" << source << std::endl;
@@ -84,7 +84,7 @@ std::string shader::log( void )
 
 	if ( res > 1 )
 	{
-		std::string info( res, 0 );
+		std::string info( static_cast<size_t>( res ), 0 );
 		glGetShaderInfoLog( _shader, res, &res, &info[0] );
 		return info;
 	}
